refactor: seperated the creation and running of the express server
